requirements.txt
Flask==3.0.3
gunicorn==23.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1